nav and reorganize
