+++ src/components/userProfile/index.js
@@ -1,0 +1,21 @@
+import React from "react";
+import "./userProfile.scss";
+
+
+function UserProfile() {
+  return (
+  
+    <div>
+      <div className="user-img" >
+        <img src="https://image.freepik.com/free-vector/businessman-character-avatar-isolated_24877-60111.jpg"
+        />
+      </div>
+      <div className="user-name" > 
+        <h4>Beezle Bryan</h4>
+      </div>
+    </div>
+ 
+  );
+}
+
+export default UserProfile;
