+++ src/components/userProfile/userProfile.scss
@@ -1,0 +1,10 @@
+.user-img {
+    display: flex;
+    width: 160px;
+    height: 160px;
+    
+    img {
+        border-radius:80px;
+        
+    }
+}
