+++ src/pages/app/App.scss
@@ -2,57 +2,3 @@
   text-align: center;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App{
-  text-align: center;
-  min-height: 100vh;
-  background-image: url("https://www.transparenttextures.com/patterns/grey-washed-wall.png");
-}
-
-.search-form{
-  min-height: 10vh;
-  display: flex;
-  justify-content: center;
-  align-content: center;
-}
-
-.search-bar{
-  width: 50%;
-  border: none;
-  padding: 10px;
-  background: #f7f9fb;
-}
-
-.search-button{
-  background: #31708e;
-  border: none;
-  padding: 10px 20px;
-  color: #f7f9fb;
-}
-
-.recipes{
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-}
