+++ src/pages/app/index.js
@@ -1,66 +1,26 @@
-import React, { useEffect, useState } from "react";
-import Recipe from "../recipes";
-
+import React from "react";
 import "./App.scss";
 
+import Nav from '../../components/nav';
+import Recipes from '../recipes';
+import Splash from '../loginSplash';
+import Users from '../user';
+
+
+import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
+
 function App() {
-  const APP_ID = "943870ca";
-  const APP_KEY = "1ea928e38226f6d40c8f3af4f45995c0";
-
-  const [recipes, setRecipes] = useState([]);
-  const [search, setSearch] = useState("");
-  const [query, setQuery] = useState("chicken");
-
-  useEffect(() => {
-    getRecipes();
-  }, [query]);
-
-  const getRecipes = async () => {
-    const response = await fetch(
-      `https://api.edamam.com/search?q=${query}&app_id=${APP_ID}&app_key=${APP_KEY}`
-    );
-    const data = await response.json();
-    console.log(data);
-    //! this console log will display the "json-ified" data for you, in order to see what hits are available.
-    setRecipes(data.hits);
-  };
-
-  const updateSearch = (e) => {
-    setSearch(e.target.value);
-  };
-
-  const getSearch = (e) => {
-    e.preventDefault();
-    setQuery(search);
-    setSearch("");
-  };
-
   return (
-    <div className="App">
-      <form onSubmit={getSearch} className="search-form">
-        <input
-          className="search-bar"
-          type="text"
-          value={search}
-          onChange={updateSearch}
-        />
-        <button className="search-button" type="submit">
-          {" "}
-          SEARCH{" "}
-        </button>
-      </form>
-      <div className="recipes">
-      {recipes.map((recipe) => (
-        <Recipe
-          key={recipe.recipe.label}
-          title={recipe.recipe.label}
-          calories={recipe.recipe.calories}
-          image={recipe.recipe.image}
-          ingredients={recipe.recipe.ingredients}
-        />
-      ))}
-    </div>
-    </div>
+    <Router> 
+  <div className="App">
+    <Nav />
+    <Switch>
+    <Route path="/" exact component={Recipes} />
+    <Route path="/splash" component={Splash} />
+    <Route path="/users" component={Users} />
+  </Switch>
+  </div>
+  </Router>
   );
 }
