+++ src/pages/loginSplash/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import "./splash.scss";
+
+
+function Splash() {
+  return (
+  <div className="Results">
+    <h1>Splash Page</h1>
+  </div>
+  );
+}
+
+export default Splash;
