+++ src/pages/recipes/index.js
@@ -1,38 +1,66 @@
-import React from "react";
-import {
-  Card,
-  Button,
-  CardImg,
-  CardTitle,
-  CardText,
-  CardGroup,
-  CardSubtitle,
-  CardBody,
-} from "reactstrap";
+import React, {useEffect, useState} from "react";
+import "./recipes.scss";
+import ResultsCard from "../../components/card";
 
-const Recipe = ({ title, calories, image, ingredients }) => {
+
+function Recipes() {
+  const APP_ID = "943870ca";
+  const APP_KEY = "1ea928e38226f6d40c8f3af4f45995c0";
+
+  const [recipes, setRecipes] = useState([]);
+  const [search, setSearch] = useState("");
+  const [query, setQuery] = useState("chicken");
+
+  useEffect(() => {
+    getRecipes();
+  }, [query]);
+  
+  const getRecipes = async () => {
+    const response = await fetch(
+      `https://api.edamam.com/search?q=${query}&app_id=${APP_ID}&app_key=${APP_KEY}`
+    );
+    const data = await response.json();
+    console.log(data);
+    setRecipes(data.hits);
+  };
+
+  const updateSearch = (e) => {
+    setSearch(e.target.value);
+  };
+
+  const getSearch = (e) => {
+    e.preventDefault();
+    setQuery(search);
+    setSearch("");
+  };
+
   return (
-    <CardGroup>
-      <Card>
-        <CardImg top width="100%" src={image} alt="Card image cap" />
-        <CardBody body inverse color="primary">
-          <CardTitle tag="h5">{title}</CardTitle>
-          <CardSubtitle tag="h6" className="mb-2 text-muted">
-            {calories}
-          </CardSubtitle>
-          <CardText>
-            <ol>
-              {" "}
-              {ingredients.map((ingredient) => (
-                <li>{ingredient.text}</li>
-              ))}
-            </ol>
-          </CardText>
-          <Button color="primary">Button</Button>
-        </CardBody>
-      </Card>
-    </CardGroup>
+    <div className="Recipes">
+   <form onSubmit={getSearch} className="search-form">
+        <input
+          className="search-bar"
+          type="text"
+          value={search}
+          onChange={updateSearch}
+          />
+        <button className="search-button" type="submit">
+          {" "}
+          SEARCH{" "}
+        </button>
+      </form>
+      <div className="recipes">
+      {recipes.map((recipe) => (
+        <ResultsCard
+        key={recipe.recipe.label}
+        title={recipe.recipe.label}
+        calories={recipe.recipe.calories}
+        image={recipe.recipe.image}
+        ingredients={recipe.recipe.ingredients}
+        />
+        ))}
+    </div>
+    </div>
   );
 };
 
-export default Recipe;
+export default Recipes;
