+++ src/pages/recipes/recipes.scss
@@ -1,0 +1,26 @@
+.search-form{
+    min-height: 10vh;
+    display: flex;
+    justify-content: center;
+    align-content: center;
+  }
+  
+  .search-bar{
+    width: 50%;
+    border: none;
+    padding: 10px;
+    background: #f7f9fb;
+  }
+  
+  .search-button{
+    background: #31708e;
+    border: none;
+    padding: 10px 20px;
+    color: #f7f9fb;
+  }
+  
+  .recipes{
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+  }
