+++ src/pages/user/index.js
@@ -1,0 +1,16 @@
+import React from "react";
+import UserProfile from "../../components/userProfile";
+import UserFavorites from "../../components/userFavorites";
+import "./user.scss";
+
+
+function Users() {
+  return (
+  <div className="Results">
+    <UserProfile />
+    <UserFavorites />
+  </div>
+  );
+}
+
+export default Users;
