+++ src/pages/user/user.scss
@@ -1,0 +1,19 @@
+.Results {
+    margin-top: 5%;
+    display: flex;
+    justify-content: space-around;
+    
+}
+
+.user-img {
+    display: flex;
+    width: 160px;
+    height: 160px;
+    
+    img {
+        border-radius:80px;
+        
+    }
+}
+
+
